add test for quickstart config
