+++ tests/acceptance/test_full_configuration.py
@@ -169,7 +169,6 @@ def test_logprep_exposes_prometheus_metrics(tmp_path):
         assert "error" not in output.lower(), "error message"
         assert "critical" not in output.lower(), "error message"
         assert "exception" not in output.lower(), "error message"
-        assert "error" not in output.lower(), "error message"
         if "Finished building pipeline" in output:
             break
     response = requests.get("http://127.0.0.1:8000", timeout=5)
